Change gitlab clone policy.
Use shallow clones of submodules, which significantly speeds up the clones,
reduces network usage as well as storage requirements.

This is also managed through the main Makefile, but using a Gitlab
feature might bring more often the benefits as we currently only support
a specific version of the Git client.