Cache the src/{panorama, {SQLP,p}lanner} folders
This is required as SBT places all the JAR files and dependencies it
fetches inside the src/<project> which is its root folder.

Unless we do this, we see failures as make will not re-trigger the build
of scala/java projects as the .*_done files are present, while the binary
files are missing as they where not kept.

Same applies for NPM.