Add package step
This is an optional step which is only executed when called manually. As
for all the other steps, it is cached and can be overriden whenever
necessary.

By default, it calls cpack to generate a .deb archive out of the current
build folder. It is up to the developper to do a debug or release build
beforehand as he wishes.